move code to namespaces
