+++ TactilityCore/Source/TactilityCore.h
@@ -6,7 +6,7 @@
 #include "CoreDefines.h"
 #include "CoreExtraDefines.h"
 #include "CoreTypes.h"
-#include "Critical.h"
+#include "critical/Critical.h"
 #include "EventFlag.h"
 #include "Kernel.h"
 #include "Log.h"
