+++ TactilityHeadless/Source/Preferences.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <cstdint>
-#include <cstddef>
 #include <string>
 
 namespace tt {
